+++ NeoConnect/ComfortLevel.cs
@@ -7,12 +7,12 @@
             if (interval != null && interval.Length >= 2)
             {
                 Time = TimeOnly.Parse(interval[0].ToString());
-                TargetTemp = decimal.Parse(interval[1].ToString());
+                TargetTemp = double.Parse(interval[1].ToString());
             }
         }
 
         public TimeOnly Time { get; private set; }
-        public decimal TargetTemp { get; private set; }
+        public double TargetTemp { get; private set; }
     }
 }
